requirements.txt
boto3==1.39.7
botocore==1.39.7
jmespath==1.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
s3transfer==0.13.0
six==1.17.0
urllib3==2.5.0
